+++ pkg/cmdserver/pkg/commands/commands_test.go
@@ -4,8 +4,8 @@ import (
 	"bytes"
 	"context"
 	"fmt"
-	"github.com/direktiv/direktiv/cmd/direktiv/cmdserver/pkg/commands"
-	"github.com/direktiv/direktiv/cmd/direktiv/cmdserver/pkg/server"
+	"github.com/direktiv/direktiv/pkg/cmdserver/pkg/commands"
+	"github.com/direktiv/direktiv/pkg/cmdserver/pkg/server"
 	"io"
 	"net"
 	"net/http"
