+++ pkg/cmdserver/pkg/server/server_test.go
@@ -12,7 +12,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/direktiv/direktiv/cmd/dirkeitv/cmdserver/pkg/server"
+	"github.com/direktiv/direktiv/pkg/cmdserver/pkg/server"
 	"github.com/stretchr/testify/assert"
 )
